# docs/openapi/phase-1.yaml
...
openapi: 3.1.0
info:
  title: UB — Phase 1 API
  description: Identity, tenancy, and catalog backbone (Phase 1).
  version: 0.0.1-SNAPSHOT
servers:
- url: /
  description: Current host
paths:
...
  /api/v1/suppliers:
    get:
      tags:
      - suppliers-controller
      operationId: list
      parameters:
      - name: pageable
        in: query
        required: true
        schema:
          $ref: '#/components/schemas/Pageable'
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PageSupplierResponse'
    post:
      tags:
      - suppliers-controller
      operationId: create
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSupplierRequest'
        required: true
      responses:
        '201':
          description: Created
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/SupplierResponse'
  /api/v1/suppliers/{supplierId}/contacts:
    get:
      tags:
      - suppliers-controller
      operationId: listContacts
      parameters:
      - name: supplierId
        in: path
        required: true
        schema:
          type: string
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SupplierContactResponse'
    post:
      tags:
      - suppliers-controller
      operationId: addContact
      parameters:
      - name: supplierId
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSupplierContactRequest'
        required: true
      responses:
        '201':
          description: Created
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/SupplierContactResponse'
...
  /api/v1/suppliers/{supplierId}:
    get:
      tags:
      - suppliers-controller
      operationId: get
      parameters:
      - name: supplierId
        in: path
        required: true
        schema:
          type: string
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/SupplierResponse'
    patch:
      tags:
      - suppliers-controller
      operationId: patch
      parameters:
      - name: supplierId
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchSupplierRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/SupplierResponse'
  /api/v1/suppliers/{supplierId}/contacts/{contactId}:
    delete:
      tags:
      - suppliers-controller
      operationId: deleteContact
      parameters:
      - name: supplierId
        in: path
        required: true
        schema:
          type: string
      - name: contactId
        in: path
        required: true
        schema:
          type: string
      responses:
        '204':
          description: No Content
    patch:
      tags:
      - suppliers-controller
      operationId: patchContact
      parameters:
      - name: supplierId
        in: path
        required: true
        schema:
          type: string
      - name: contactId
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchSupplierContactRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/SupplierContactResponse'
...
components:
  schemas:
...
    CreateSupplierRequest:
      type: object
      properties:
        name:
          type: string
          maxLength: 255
          minLength: 0
        code:
          type: string
          maxLength: 64
          minLength: 0
        supplierType:
          type: string
          maxLength: 32
          minLength: 0
        vatPin:
          type: string
          maxLength: 64
          minLength: 0
        taxExempt:
          type: boolean
        creditTermsDays:
          type: integer
          format: int32
        creditLimit:
          type: number
        status:
          type: string
          maxLength: 16
          minLength: 0
        notes:
          type: string
          maxLength: 5000
          minLength: 0
        paymentMethodPreferred:
          type: string
          maxLength: 32
          minLength: 0
        paymentDetails:
          type: string
          maxLength: 2000
          minLength: 0
      required:
      - name
    SupplierResponse:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        code:
          type: string
        supplierType:
          type: string
        vatPin:
          type: string
        taxExempt:
          type: boolean
        creditTermsDays:
          type: integer
          format: int32
        creditLimit:
          type: number
        rating:
          type: number
        status:
          type: string
        notes:
          type: string
        paymentMethodPreferred:
          type: string
        paymentDetails:
          type: string
        version:
          type: integer
          format: int64
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateSupplierContactRequest:
      type: object
      properties:
        name:
          type: string
          maxLength: 255
          minLength: 0
        roleLabel:
          type: string
          maxLength: 128
          minLength: 0
        phone:
          type: string
          maxLength: 64
          minLength: 0
        email:
          type: string
          maxLength: 255
          minLength: 0
        primaryContact:
          type: boolean
    SupplierContactResponse:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        roleLabel:
          type: string
        phone:
          type: string
        email:
          type: string
        primaryContact:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    PatchSupplierRequest:
      type: object
      properties:
        name:
          type: string
          maxLength: 255
          minLength: 0
        code:
          type: string
          maxLength: 64
          minLength: 0
        supplierType:
          type: string
          maxLength: 32
          minLength: 0
        vatPin:
          type: string
          maxLength: 64
          minLength: 0
        taxExempt:
          type: boolean
        creditTermsDays:
          type: integer
          format: int32
        creditLimit:
          type: number
        status:
          type: string
          maxLength: 16
          minLength: 0
        notes:
          type: string
          maxLength: 5000
          minLength: 0
        paymentMethodPreferred:
          type: string
          maxLength: 32
          minLength: 0
        paymentDetails:
          type: string
          maxLength: 2000
          minLength: 0
    PatchSupplierContactRequest:
      type: object
      properties:
        name:
          type: string
          maxLength: 255
          minLength: 0
        roleLabel:
          type: string
          maxLength: 128
          minLength: 0
        phone:
          type: string
          maxLength: 64
          minLength: 0
        email:
          type: string
          maxLength: 255
          minLength: 0
        primaryContact:
          type: boolean
...
    Pageable:
      type: object
      properties:
        page:
          type: integer
          format: int32
          minimum: 0
        size:
          type: integer
          format: int32
          minimum: 1
        sort:
          type: array
          items:
            type: string
...
    PageableObject:
      type: object
      properties:
        offset:
          type: integer
          format: int64
        sort:
          $ref: '#/components/schemas/SortObject'
        paged:
          type: boolean
        pageNumber:
          type: integer
          format: int32
        pageSize:
          type: integer
          format: int32
        unpaged:
          type: boolean
    SortObject:
      type: object
      properties:
        empty:
          type: boolean
        sorted:
          type: boolean
        unsorted:
          type: boolean
    PageSupplierResponse:
      type: object
      properties:
        totalPages:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        size:
          type: integer
          format: int32
        content:
          type: array
          items:
            $ref: '#/components/schemas/SupplierResponse'
        number:
          type: integer
          format: int32
        sort:
          $ref: '#/components/schemas/SortObject'
        first:
          type: boolean
        last:
          type: boolean
        numberOfElements:
          type: integer
          format: int32
        pageable:
          $ref: '#/components/schemas/PageableObject'
        empty:
          type: boolean
...
  securitySchemes:
    bearer:
      type: http
      description: 'Slice 3: access JWT via Authorization header.'
      scheme: bearer
      bearerFormat: JWT
    apiKey:
      type: apiKey
      description: 'Slice 3: tenant integration key (prefix + secret; returned once
        at create).'
      name: X-API-Key
      in: header
    cookieRefresh:
      type: apiKey
      description: Optional HttpOnly refresh cookie; native clients may use JSON body
        instead.
      name: ub_refresh
      in: cookie